+++ src/main/java/com/grepp/spring/app/model/quiz/dto/request/SurvivalResultRequest.java
@@ -1,4 +1,4 @@
-package com.grepp.spring.app.controller.api.quiz.payload;
+package com.grepp.spring.app.model.quiz.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
