+++ src/main/java/com/grepp/spring/app/model/quiz/dto/response/QuizRegisterResponse.java
@@ -1,4 +1,4 @@
-package com.grepp.spring.app.controller.api.quiz.payload;
+package com.grepp.spring.app.model.quiz.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
