+++ src/main/java/com/grepp/spring/app/model/quiz/repository/quizSetRepository/QuizSetRepositoryCustomImpl.java
@@ -1,6 +1,6 @@
 package com.grepp.spring.app.model.quiz.repository.quizSetRepository;
 
-import com.grepp.spring.app.model.quiz.dto.QuizProjection;
+import com.grepp.spring.app.model.quiz.dto.maybeResponse.QuizProjection;
 import com.querydsl.core.types.Projections;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
